fix: prevent window resizing when on submit/receive/open/close
This behaviour can get annoying: when any action restores the side panel
to its original size eventhough the user might have chose to re-size the
panel. I’m having to measure the width of the window to restore it after
updates. Not pleasant, but it’s the best I have atm.